Release v1.4
This may be the final update that Defenestrate receives. Everything that was planned is now finished. If I think of some other feature to add or some other cool feature is requested, I'll definitely make another update for that, but I can finally say that this project is finished.

Changelog:
- made missing configuration values load defaults
- added option to disallow players from stealing entities and blocks off of others' heads
- this plugin now supports picking up and throwing containers without the items within disappearing
- added block particles to blocks that are being picked up